Display filtered reserved rockets and style my-profile component
