--- src/components/Dragons.js
@@ -1,3 +1,0 @@
-const Dragons = () => <h1>Dragons coming soon...</h1>;
-
-export default Dragons;
