--- src/components/Profile.js
@@ -1,3 +1,0 @@
-const Profile = () => <h1>Profile coming soon... </h1>;
-
-export default Profile;
