--- src/components/Rockets.js
@@ -1,3 +1,0 @@
-const Rockets = () => <h1>Rockets coming soon...</h1>;
-
-export default Rockets;
