+++ src/pages/Profile.js
@@ -1,39 +1,39 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-// import { getRocketsAction } from '../store/reducers/actions/actions';
+import { getRocketsAction } from '../store/reducers/actions/rocketActions';
 import { fetchData } from '../store/reducers/actions/dragonActions';
 
-// const Profile = () => {
-//   const { rockets } = useSelector((state) => state.rockets);
+export const RocketProfile = () => {
+  const dispatch = useDispatch();
+  const { rockets } = useSelector((state) => state.rockets);
 
-//   useEffect(() => {
-//     if (!rockets) {
-//       dispatch(getRocketsAction());
-//     }
-//   }, []);
+  useEffect(() => {
+    if (!rockets) {
+      dispatch(getRocketsAction());
+    }
+  }, []);
 
-//   return (
-//     <section className="my-profile inner-padding">
-//       <div>
-//         <h1>My Missions</h1>
-//       </div>
-//       <div className="rockets">
-//         <h1>My Rockets</h1>
-//         <ul className="list-container">
-//           {rockets
-//             && rockets.filter((rocket) => (
-//               <li key={rocket.id} className="list-item">
-//                 {rocket.rocket_name}
-//               </li>
-//             ))}
-//         </ul>
-//       </div>
-//     </section>
-//   );
-// };
-// export default Profile;
+  return (
+    <section className="my-profile inner-padding">
+      <div>
+        <h1>My Missions</h1>
+      </div>
+      <div className="rockets">
+        <h1>My Rockets</h1>
+        <ul className="list-container">
+          {rockets
+            && rockets.filter((rocket) => (
+              <li key={rocket.id} className="list-item">
+                {rocket.rocket_name}
+              </li>
+            ))}
+        </ul>
+      </div>
+    </section>
+  );
+};
 
-const Profile = () => {
+export const DragonProfile = () => {
   const dispatch = useDispatch();
   const dragons = useSelector((state) => state.dragons.dragons);
 
@@ -62,4 +62,3 @@ const Profile = () => {
     </div>
   );
 };
-export default Profile;
