+++ src/pages/Rockets.js
@@ -4,16 +4,16 @@ import {
   cancelReservationAction,
   getRocketsAction,
   setReserveAction,
-} from '../store/reducers/actions/actions';
+} from '../store/reducers/actions/rocketActions';
 
 const Rockets = () => {
-  const dispatch = useDispatch();
   const { rockets } = useSelector((state) => state.rockets);
-  const loadRockets = () => {
-    dispatch(getRocketsAction());
-  };
+  const dispatch = useDispatch();
+
   useEffect(() => {
-    loadRockets();
+    if (!rockets) {
+      dispatch(getRocketsAction());
+    }
   }, []);
 
   const setReserve = (id) => {
@@ -27,7 +27,7 @@ const Rockets = () => {
   return (
     <section className="inner-padding">
       {rockets
-        && rockets.filter((rocket) => (
+        && rockets.map((rocket) => (
           <div className="rocket" key={rocket.id}>
             <div className="rocket-img">
               <img src={rocket.flickr_images} alt="rocket" />
@@ -35,9 +35,7 @@ const Rockets = () => {
             <div className="rocket-info">
               <h1>{rocket.rocket_name}</h1>
               <p>
-                {rocket.reserved && (
-                  <span className="reserve-badge">Reserved</span>
-                )}
+                {rocket.reserved && <span className="reserve-badge">Reserved</span>}
                 {rocket.description}
               </p>
               <div className="button-div">
